refactor: move vault utility
