+++ src/common/storage.ts
@@ -1,3 +1,5 @@
+export const hasSetPasswordIdentifier = 'stacks-wallet-has-set-password' as const;
+
 export enum StorageKey {
   'authenticationRequests',
   'transactionRequests',
@@ -60,4 +62,12 @@ export function getRequestOrigin(storageKey: StorageKey, request: string): strin
 export function deleteTabForRequest(storageKey: StorageKey, request: string) {
   const key = getKeyForRequest(storageKey, request);
   localStorage.removeItem(key);
+}
+
+export function getHasSetPassword() {
+  const persisted = localStorage.getItem(hasSetPasswordIdentifier);
+  if (persisted !== null) {
+    return JSON.parse(persisted);
+  }
+  return false;
 }
